Auth template

+++ src/upvoteThread/upvoteThread.controller.ts
@@ -1,4 +1,5 @@
 import * as upvoteModel from './upvoteThread.model';
+import * as authModel from '../auth/auth.model';
 import { Request, Response } from 'express';
 
 export async function index(req: Request, res: Response) {
@@ -22,30 +23,52 @@ export async function index(req: Request, res: Response) {
 
 export async function upvote(req: Request, res: Response) {
   try {
-    const username = req.params.username;
-    const projectId = parseInt(req.params.projectId);
-    const threadId = parseInt(req.params.threadId);
-    const payload = {
-      user_name: username,
-      project_id: projectId,
-      thread_id: threadId,
-    };
-    await upvoteModel.create(payload);
+    const cookieObj: { sessionToken: string } = req.cookies;
+    const sessionId: string = cookieObj.sessionToken;
+    const ghuid = await authModel.validateUser(sessionId);
 
-    res.status(200).send('');
+    if (!ghuid) {
+      return res.status(404).send('Invalid Access Token');
+    }
+    try {
+      const username = req.params.username;
+      const projectId = parseInt(req.params.projectId);
+      const threadId = parseInt(req.params.threadId);
+      const payload = {
+        user_name: username,
+        project_id: projectId,
+        thread_id: threadId,
+      };
+      await upvoteModel.create(payload);
+
+      res.status(200).send('');
+    } catch (error: any) {
+      res.status(500).send(error.message);
+    }
   } catch (error: any) {
-    res.status(500).send(error.message);
+    res.status(404).send(error.message);
   }
 }
 
 export async function remove(req: Request, res: Response) {
   try {
-    const username = req.params.username;
-    const threadId = parseInt(req.params.threadId);
-    await upvoteModel.deleteById(username, threadId);
+    const cookieObj: { sessionToken: string } = req.cookies;
+    const sessionId: string = cookieObj.sessionToken;
+    const ghuid = await authModel.validateUser(sessionId);
 
-    res.status(200).send('');
+    if (!ghuid) {
+      return res.status(404).send('Invalid Access Token');
+    }
+    try {
+      const username = req.params.username;
+      const threadId = parseInt(req.params.threadId);
+      await upvoteModel.deleteById(username, threadId);
+
+      res.status(200).send('');
+    } catch (error: any) {
+      res.status(500).send(error.message);
+    }
   } catch (error: any) {
-    res.status(500).send(error.message);
+    res.status(404).send(error.message);
   }
 }
